user-can-delete-items-from-cart
