+++ client/src/pages/ProductDetails.js
@@ -1,36 +1,12 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import './ProductDetails.css'
+import { addToCart } from './Shopping-cart.js'
 
 async function fetchProduct(productId) {
   const res = await fetch((`/api/products/${productId}`));
   if (!res.ok) throw new Error(`fetch Error ${res.status}`);
   return await res.json();
-}
-
-// create cart file
-// function for get cart contents
-// function for add , and another for remove
-
-function getCart() {
-  let cart = JSON.parse(localStorage.getItem('cart'));
-  if (cart === null) {
-    return [];
-  } else return cart
-}
-
-function addToCart(product) {
-  const cart = getCart()
-  cart.push(product);
-  localStorage.setItem('cart', JSON.stringify(cart))
-  window.location.reload()
-}
-
-export function removeFromCart(e) {
-  console.log('hello')
-  // console.log(e.Target.productId);
-  localStorage.clear()
-  window.location.reload()
 }
 
 export default function ProductDetails() {
@@ -68,36 +44,39 @@ export default function ProductDetails() {
   return (
     <div className="container prod-details-container">
       <div className="row">
-          <div className="col-12 col-md-6 col-lg-6">
-              <img src={url} alt={name} className="image" />
-          </div>
-          <div className="col-12 col-md-6 col-lg-6">
-            <div>
-              <h5>{name}</h5>
-              <p>{color}</p>
-              <p>{description}</p>
-              <ul>
-                <li>Relaxed fit</li>
-                <li>High wicking and quick drying fabrics</li>
-                <li>Fully-dyed fabric</li>
-                <li>Guarded zip ends</li>
-                <li>Zippered side pocket with protected insert</li>
-                <li>Robust YKK zipper with semi auto-lock puller</li>
-                <li>Fabric: 86% Polyester, 14% Elastane</li>
-              </ul>
-              <div className='button-div'>
-                <a
+        <div className="col-12 col-md-6 col-lg-6">
+          <img src={url} alt={name} className="image" />
+        </div>
+        <div className="col-12 col-md-6 col-lg-6">
+          <div>
+            <h5>{name}</h5>
+            <p>{color}</p>
+            <p>{description}</p>
+            <ul>
+              <li>Relaxed fit</li>
+              <li>High wicking and quick drying fabrics</li>
+              <li>Fully-dyed fabric</li>
+              <li>Guarded zip ends</li>
+              <li>Zippered side pocket with protected insert</li>
+              <li>Robust YKK zipper with semi auto-lock puller</li>
+              <li>Fabric: 86% Polyester, 14% Elastane</li>
+            </ul>
+            <div className='button-div'>
+              <a
                 href='http://localhost:3000/shoppingcart'
-                onClick={function () {
-                    addToCart(product);
-                  }
-                  }>
-                  ADD TO CART
-                </a>
-              </div>
+                onClick={() => addToCart(product)}
+                >
+                ADD TO CART
+              </a>
             </div>
           </div>
+        </div>
       </div>
     </div>
   );
 }
+
+//  to be moved ///////////////////////
+// create cart file
+// function for get cart contents
+// function for add , and another for remove
