+++ client/src/pages/Shopping-cart.js
@@ -1,8 +1,7 @@
 import './Shopping-cart.css'
-import { removeFromCart } from './ProductDetails';
+import CartItems from '../components/CartItems.js';
 
 export default function ShoppingCart() {
-  if (items) {
     return (
       <div className="container prod-details-container cart-items-container">
         <h1 className='cart-header'>Cart</h1>
@@ -29,13 +28,16 @@ export default function ShoppingCart() {
 
       </div>
     )
-  } else return <h1 className='cart-header'>Your cart is empty</h1>
-
 }
 
-const items = JSON.parse(localStorage.getItem('cart'));
-console.log(items)
+// grab all items from local storage //
+export function localStorageCart() {
+  const items = JSON.parse(localStorage.getItem('cart'));
+  return items
+}
+const items = localStorageCart();
 
+// total all items in cart //
 function subtotal() {
   let sum = 0;
   if (items) {
@@ -47,34 +49,24 @@ function subtotal() {
 }
 const amount = subtotal()
 
-function CartItems() {
-   const cartItem = items.map(x =>
-    <div key={x.productId} className = "row" >
-      <div className="col-6 col-md-6 col-lg-6">
-        <img src={x.url} alt='jersey' className="cart-image" />
-      </div>
-      <div className="col-6 col-md-6 col-lg-6 cart-item-text">
-        <div>
-          <div className='name-and-price'>
-            <p>{x.name}</p>
-            <p className='dollar-amt'>${x.price}</p>
-          </div>
-          <p>{x.color}</p>
-          <div className='remove-button-div'>
-            <a
-               href='http://localhost:3000/'
-            onClick={removeFromCart}
-            >
-              Remove All Items
-              </a>
-          </div>
-        </div>
-      </div>
-    </div >
-    )
-  return (
-    <div>
-      {cartItem}
-    </div>
-  )
+// functions for manipulating cart //
+export function getCart() {
+  let cart = JSON.parse(localStorage.getItem('cart'));
+  if (cart === null) {
+    return [];
+  } else return cart
+}
+
+export function addToCart(product) {
+  const cart = getCart()
+  cart.push(product);
+  localStorage.setItem('cart', JSON.stringify(cart))
+  window.location.reload()
+}
+
+export function removeFromCart(productId) {
+  let cart = JSON.parse(localStorage.getItem('cart'));
+  const newCart = cart.filter((item) => item.productId !== productId)
+  localStorage.setItem('cart', JSON.stringify(newCart))
+  window.location.reload()
 }
